move experiments/* to lagrangebench/runner.py
