+++ configs/ldc_2d/segnn.yaml
@@ -13,7 +13,7 @@ optimizer:
   noise_std: 0.001
 
 logging:
-  wandb_project: ldc_3d
+  wandb_project: ldc_2d
 
 neighbors:
   multiplier: 2.0
