+++ configs/rpf_2d/segnn.yaml
@@ -12,4 +12,4 @@ optimizer:
   lr_start: 1.e-3
 
 logging:
-  wandb_project: rpf_3d
+  wandb_project: rpf_2d
